vmbackup: add null provider, do late initialization of sync provider.
This change adds a null provider so that systems that don't have
either VSS nor the sync driver (or have them disabled) can still execute
at least backup scripts when taking a quiesced snapshot. This should
bring feature parity to the previous quiesced snapshot approach, which
would use the VIX path if the vmbackup protocol wasn't supported by
the Tools.

Also initialize the sync provider late, only when starting a backup
operation. This allows the users to change the Tools configuration
and not have to restart the Tools service to get the new settings,
making it easier to switch between VSS / sync driver / null provider.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>