[RC Fix] Fix bug 548177.
When user deletes a share on the host, nautilus (ubuntu file manager)
does not like us sending -EIO on a share. We detect this error and map
it to success and create minimal fake attributes.

In future we this client will this mask new error codes that we add to
the server as well, which is sad.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>