Fail more gracefully on unrecognized I18N encodings
Unicode_Init has a check for unsupported local character encodings,
but the code that prints the unsupported encoding name calls
Unicode_EncodingEnumToName(), which assumes that the enum passed
into it is valid.  Since the encoding name wasn't found, the enum
is STRING_ENCODING_UNKNOWN, which results in a not particularly
graceful or informative NOT_REACHED() down inside lib/unicode.
The fix is to use the string that codeset gave us instead of
trying to convert STRING_ENCODING_UNKNOWN back into an encoding
name.  Also add an ASSERT to Unicode_EncodingEnumToName() to
check that STRING_ENCODING_UNKNOWN is not passed to it.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>