multi thread fix for evct and fcvt use in bsd printf
Can't use mxuser locks for this because they'll pull in lib/lock
all over the place.  On Windows use the thread-safe functions _ecvt_s
and _fcvt_s instead of having to dynamically allocate a mutex.
There are thread-safe versions for Linux, too, ecvt_r and fcvt_r,
but they are not available on Mac.  Since there was going to be a
mutex in this code for Mac, might as well use it for Linux, too.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>