Refactor toolbox-cmd command handling.
This is mostly a mechanical change to have the sources better organized.
The change moves the command / help handlers to the command-specific
source files, so that there is a clear interface between the main()
function and the command handlers, and also so that it's less messy to
add new commands.

On top of that, a couple of minor issues (missing parameters to vararg
calls) were fixed, and the translations were updated to match the new
help messages (which have the command as a parameter, instead of hardocded
in the message).

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>