lib/lock: make the userland portion available for tools
We're using MXUser (userland) locks all over bora/lib. Some of these
libraries are showing up in the open source tools build. Make only
the MXUser portion of lib/lock available to the tools. We're OK with
open sourcing this.

Once this change goes in bora/lib changes that depend on it will
start showing up almost immediately.

I've elected to a a minimumally invasive approach to ulRec.c, one
that optimized out the conditional for acquisition and release
(by far the largest usage) and leaving the farless used TryLock
and IsLocked routines using the same conditional as before this
change.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>