Add auto-upgrade command to toolbox CLI.
This change adds a new "upgrade" command to the toolbox CLI on Win32 and
Linux (non-open-vm-tools), which are the only platforms where we have
auto-upgrade.

Add a few utility functions to the common toolbox CLI code to do a few
useful things:
. print messages based on the status of the "quiet" flag.
. send RPCs to the host using the RpcChannel API.

On top of that, put some common definitions of commands and replies in a
shared header to avoid having duplicated strings in different parts of the
source code.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>