lib/lock: press for a harder MX/MXUser separation
Since there is a function for MX to register hook functions with
MXUser, enhance it to provide additional function that allow
removing even more MX references throughout MXUser.

This change makes it possible to expose the bind MX to MXUser
functionality to userland - we can use that functionality outside
the VMX now.

The mutex.h and userlock.h include files are tweaked so that things
work properly if either of them comes first and the userlock.h file
is able to stand on its own inside and outside the VMX.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>