Don't crash when reconfiguring file log handler.
The log code has a log statement that prints info about the core dump
config; when using the file handler, and swithing between "no logging"
and "log enabled", the new log data would then open the log file; later,
when restoring the previous log config, that would cause an assert.

Fix is easy: just change the order in which those things are done.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>