lib/misc: create a hostinfo time base with ns units
Provide nanosecond reporting routines; microseconds are getting a bit
"long in the tooth" for some timing applications. Refactor the
existing code to have the lowest overhead possible based on the
platform. Obtain the best we can from the existing clock sources.

The quest for better timer sources will come next...

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>