Thread fix for TimeUtil_DaysAdd
TimeUtil_DaysAdd was not thread safe because it kept the number
of days in each of the 12 months in a static array, and set the
value for February based on whether the function was operating
on a common year or a leap year.  A simple fix is to keep two
arrays, one for common years and another for leap years.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>