Fix error on creating new files/folders in root share
When a user or inadvertently tries to create a new file or folder
or overwrite a folder in the root share fail with the correct
error. Previously the Windows hgfs client always failed with
name not found which is not the correct error and confuses some
applications which then mislead the user.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>