least and most significant bit set functions: support !x86*
Since the chances of not using a GNU compiler for ARM or any
other architecture we'll use is very small (we know GNU well
and "like" it), code the least and most significant bit set
functions such that they should work "anywhere" via the GCC
intrinsic functions.

This is a cautious beginning to possibily using the instrinics
outside the monitor and vmkernel and possibly within. For now
the X86* support is left unmodified.

Many experiments will soon follow this change...

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>