Add MSI/MSI-X support to VMCI
This change adds MSI/MSI-X support to VMCI.

The device is now MSI/MSI-X capable, and supports up to 2
vectors in MSI-X mode, one for datagrams and another for
the notification bitmap.

The Windows and Linux drivers have been updated to use
MSI/MSI-X when running on the new device.  The drivers only
support one vector for now, for datagrams.

Verified on Windows guests by checking the Device Resources.
MSI-X is selected on Win7, while on Win2K, it correctly
falls back to legacy mode.  The old Windows driver runs
correctly on the new device.

Verified on Linux by checking lspci.  MSI-X is selected on
newer kernels where CONFIG_PCI_MSI is defined, otherwise
it falls back to legacy mode.  The old Linux driver runs
correctly on the new device.

Verified on v7 and v8 hardware.  On v7, the new driver falls
back to legacy mode.  Suspending on v7 and resuming on v8
works correctly, it will use legacy and continue to fire
interrupts as expected.  On v8, it will use MSI/MSI-X
if possible.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>