Remove VIX_COMMAND_[GET|SET|]_PROPERTY.
They are unused. The Tools does process GET_PROPERTY like
GET_TOOLS_STATE, but I couldn't find or remember a version of the
client library that sent a command with this op code for this
purpose.
Also, as a side benefit, remove the flag
VIX_COMMAND_GUEST_RETURNS_ENCODED_STRING, since no request message
sets this flag, and it is a left over remenant from the incorrect
assumption that GuestRPC couldn't handle binary data. This gives
an additional benefit since inspection suggests that there's a bug
in the VMX code for processing responses when this flag is set:
the response length is computed using the encoded string, not
the decoded string (causing random VMX memory to be sent in the
response).

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>