Wiper: recognize HFS partitions, also FAT partitions mounted as 'msdos'
A user on the forums was asking about shrinking disks in Mac OS guests.
The command "vmware-tools-cli disk shrink /" currently shows an error:
"Partition / is not shrinkable".

This is because the wiper library does not recognize the mounted HFS
partitions as wipeable. This change recognizes the "hfs" type as a new
PARTITION_HFS value to support shrinking HFS partitions. It also maps
the "msdos" mount type to PARTITION_FAT to support shrinking FAT/FAT32
partitions on Mac OS guests.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>