Change how topology is changed with the new vmwgfx X driver
Multimon has always been a bit tricky to get to stick when running
a gnome-session since after we turned on multimon gnome would turn the
new output off.

This along with changes to the kernel driver and X driver fixes this.

So now the driver exports the vmware ctrl protocol and RandR1.2 and
expects vmware-user to first the set the topology via the vmware ctrl
and then setup the correct layout via RandR1.2. This is closer to what
the old path did but instead of just setting a single mode that covers
the whole topology it setups the layout as well.

The changes to the kernel driver and X driver will follow.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>