lib/lock: got native
Remove all usage of the VThreadID data type from MXUser. This
includes calling VThread_CurID. This prevents a host of chicken
before the egg problems that are wrapped around vthreadBase.c.

In order to do this two APIs had to change:

1) The per thread contention statistics had to be removed.

This is a loss since there are no tools that were interested
in the data.

2) MXUser_AnyLocksHeld

Changed so a VThreadID wasn't coming in. Trivial. It's
MXUser_IsCurThreadHoldingLocks now.

The check across everything thread case isn't necessary.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>