Rank the lib/log lock
This was a toughie. The problem here is that everyone calls Log(), Warning(), and the like while holding their own locks, yet LogWriteLogFile() calls a great deal of code itself while processing those calls. Some of this code is high-level and either already takes its own lock or is scheduled to get locking.

We considered several radical solutions, but none of them seemed satisfactory. We could have LogWriteLogFile() simply dump messages on a leaf-lock-protected queue and return, but then we'd have to create a lazy-writer logging thread. We could have LogWriteLogFile() not do complicated stuff like log file rotation itself, but schedule that for later, but again we'd have to invent a lazy mechanism (either a thread or callbacks provided by the log module initializer). And so on.

I settled on a practical solution: coming up with criteria for code that can be called underneath the log lock, and then moving around code in log.c to make it so. The criteria, for now, are quite simple - no code that LogWriteLogFile() calls under the log lock may take its own lock in turn. Basically this means stick with lib/misc and lib/file.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>