Add HashTable_FreeUnsafe to allow freeing of an atomic hash table
The MXUser read-write lock wants to free its hash table when the lcok
is destroyed, but HashTable_Free has an ASSERT that it cannot be
atomic.  Let's add another function HashTable_FreeUnsafe to allow
freeing of an atomic hash table, but the caller has to ensure that no
other thread may be accessing the hash table when it is being freed.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>