Show folder icon when dragging directories
When dragging a folder from the guest to the host we would show the wrong icon next to the cursor.

The problem was that there was no way to tell if dragged object was a file or a folder. Fix was to incldue file attributes with the drag information.

Note, this change only implements file attributes for Windows tools.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>