Make vmware-user an empty application.
This change removes the code from vmware-user, allowing it to build,
so that the installer builds still work. This allows changes in the
shared libraries (such as lib/dndGuest and lib/unity) to be made to
support the new plugins, before switching the installer to consume
them.

Drop (for now at least) the Unity features from the open-vm-tools package.

Stop building vixTools as a standalone library.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>