lib/file: preparatory change toward simplifying FileLock_Unlock
The FileLock_Lock routine returns an abstract file lock token
representing the lock on the file. There is no functional need
to base the name of the file to be unlocked to FileLock_Unlock
since the token should already contain the necessary information
for the unlock.

This change lays the groundwork. The next one will clean up all
of the FileLock_Unlock callers. In the end all of this is being
done because "it's the right thing to do" and because Kevin has
a need for it - it will make a mainMem cleanup easier.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>