lib/lock: break out your Kevlar
This was pointed out in comments from the original barrier review; I had
a "bullet proof barrier" change coming. Here you go.

Barriers are often used like this:

barrier
code
barrier

If the code is small enough or the code has a variable amount of
work and/or there are scheduler "funnies" it is possible to have
a thread arrive at the second barrier while threads are still exiting
the first barrier. That can lead to some very unobvious and strange
bugs.

The original barrier would detect this and assert. Let's do better.

This version catches the threads entering and parks them separate from
the threads leaving ensuring the everything works as expected; no
violation of the principal of least surprise.

This adds little in the way of complexity and memory use while
providing a great deal of protection (~type IIa).

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>