lib/lock: move the stats data under a single pointer
Each of the MXUser object that collects statistics has its
statistical data ifdef compiled into its internal structure. Move
the statistical data into a private structure accessed from a
single pointer with the private structure.

This cleans up a lot of MXUSER_STATS ifdefs which become unnecessary
and get the code very close to being able to runtime turn on of
statistics in any build. For now, statistics are for stats builds only.

All this is done without introducing any overhead to speak of, only
one memory read from a cache line already loaded.

This is also a prepratory change to fix a problem in Log - the lock
allocation cannot call anything that might Log.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>