lib/log: the log lock has magical properties
The log facility must use a lock. If the lock acquisition routine
were to use Log/Warning an infinite recursion is set up:

Log -> Lock -> Log -> Lock ...

It makes it imperitive that the log facility lock never have the
potential to emit a message. We just don't know where a developer
will place a Log/Warning.

The fix for this is an alternative lock creation routine, limited
to the recursive lock required for the log facility, that allows
the specification of a new parameter - beSilent (damn it!).

Ultimately this works because the core locking logic is inherently
silent; only the statistics code, if present, has to be bypassed.

Since this ultimately fixes a PR, the work-around to this issue
can also be removed.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>