Remove vmware-user from open-vm-tools.
This change does a little bit more than just remove the sources. It also
modifies the vmware-user-suid-wrapper so that it executes
"vmtoolsd -n vmusr" in open-vm-tools. The desktop file is also fixed to
execute the suid wrapper, which is the right thing to do.

Also remove locationsdb.c from open-vm-tools, since it's not used there.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>