lib/log: Formally handle the maximum message length
Provide a way to ask what the maximum message length is. Use it
to avoid creating a line too long within the MXUser lock statistics
code.

With a bit of minor code rearranging, only a single, full-sized
message buffer needs to be kept on the stack to log a message.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>