Fix a possible buffer overrun. Remove an extra space character.
We want argBufPtr to point to the end of the argBuf buffer, where the '\0'
character is. So argBufPtr should be set to  argBuf +  strlen(argBuf).
Otherwise, if the argument string fully fills up argBuf, we could have
a buffer overrun when *argbufPtr = '\0' is performed.

The other change removes the extra space character at the end of
cmdLine in the ExtractCommandLineFromAddressSpaceFile function.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>