PollDefault: Log lock data before ASSERT to aid debugging
Add a function to print out an MXUserRecLock's internal data, and use
that before an ASSERT in PollDefault that checks for real-time
callback that is starved by a lock.  This is necessary as for some
reason the hostd cores that were generated due to this ASSERT never
have the state of the lock and the lock owner at the time of the ASSERT
(the lock is always free in the core and the thread that probably held
the lock is not anywhere such lock is held).

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>