VMCIDoorbell_Destroy: Assert when destroying non existent handle.
The VMCINotificationUnregister function in the vmci guest driver
was asserting FALSE when destroying a doorbell handle that does not
exist in the internal hash. This was causing the guest kernel to panic
when trying to destroy a handle that was already destroyed.

We should probably just return VMCI_ERROR_NOT_FOUND here instead of
asserting.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>