Fix renames on non-Windows hosts
The check if the target file exists and the rename was not to
replace an existing target was badly coded.
The logic left the status field incorrectly set to file not found
which was fine for not replacing the target but then the rename
was missed due to a check on the status being successful for the
existance check.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>