Handle resourcePoolPath overflow situation (host side).
We enforce the maximum size of resourcePoolPath on both ends of guestlib wire
protocol (both versions 2 and 3). We handle larger buffers by not sending
them in the existing protocols.

This change is also a minor revision to the guestlib v3 protocol (host side
only) to add another field to the v3 layout to send the larger
resourcePoolPath buffer as a byte array. The guest client side is omitted
here.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>