Add a monotonic glib timer to Tools.
While investigating bug 600636, I noticed that glib timers sometimes would
get into a funny state where they'd ask the main loop to sleep for a very
short time, even though they weren't ready to be fired. I couldn't figure
out why they were doing that, but anyway, we can't easily modify glib.

Instead, add a new timer, based on a monotonic clock, to use in preference
to the glib timer. This is more important for quick firing timers such as
the RPC loop or the C&P pointer poll than it is for coarser-grained timers
such as the guest info gather loop.

Along with it, change the rpcin library and the dnd plugin to use the new
timer in place of the glib one, and avoid re-creating the timer on every
iteration of the rpcin loop - do it only when the delay has changed.

I was only able to reproduce the original issue on a 32-bit Win XP VM. I
tried a 64-bit Win 7 VM and a 64-bit Linux VM, and both seemed to behave
fine.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>