Only enter gtk_main_iteration when Tools app main loop is active.
If vmtoolsd mainloop is, or becomes, inactive, we need to stop issuing
pointer grab clipboard checks. It was advised the way to do this was
to look at the main loop field of the tools app context and see if it
is active, and only enter gtk_main_iteration() if so.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>