Unregister legacy backdoor in X11 UI class destructor.
The legacy backdoor copy paste (used by ESX) is registered in
an Init function of the X11 UI class. We should only unregister
the legacy copy paste when this class is destructed.

Up to now, were were doing it as a side effect of resetting
dndcp version 3 and higher in response to a reset signal from
the plugin framework. This would cause the backdoor copypaste
code to unregister at reset without being registered once again
because the registration only happened in Init.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>