Implement Windows hibernation support
This change improves the VMCI Windows guest drivers handling
of hibernation. Before this change, any device state that is
partly or fully mirrored in the device driver as well, would
become out of sync when coming out of hibernation, e.g., queue
pairs would no longer be allocated on the hypervisor side but
the guest driver would still believe that they were active.

This change deals with such state in two ways: 1) For queue
pairs, all non-local queue pairs are converted to local ones
when going into hibernation. This part utilizes the exisiting
support on the Windows host side for converting queue pairs to
local memory (on the Windows host side it is used when the VMX
goes away). 2) For doorbells, we simply reregister any
doorbells when leaving hibernation.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>