Retire VMCIQPair_Init
Queue pairs allocated with VMCIQPair_Alloc are initialized by
either the hypervisor or the VMCI driver (for local queue
pairs) and has been doing so since the initial release of VMCI
with queue pairs (ws65), so VMCIQPair_Init isn't necessary
(and may not do what you think, since it only initializes half
a queue pair). So retire it.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>