guest info: ignore interfaces with invalid MAC address.
When the OS is messing with interfaces (such as bringing them up or down),
the guest info plugin may see an interface with an invalid MAC address, at
least on Solaris. Trying to write that iface's data to the nic info
structure causes the plugin to crash (probably because addr_ntoa doesn't
understand the address and is returning garbage, that we then try to read
when writing it as a string).

So just skip those. When the gather loop runs again, we'll pick up any
changed interfaces.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>