Incorrect error code while sending dgram in nonblock mode to host.
Actually this error was occuring on both ends.  We were incorrectly
mapping VMCI_ERROR_INVALID_RESOURCE to EINVAL, which is wrong.
It really means we cannot find the destination endpoint.  Fixed
the mapping.

Also discovered a bug in the Linux driver, where we would hit an
ASSERT when the length of the buffer passed to recv() was zero.
This is a valid case on Linux, recv on a zero-length should pass,
so fixed that too.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>