Check for XDR limits when populating guest info structs.
We defined some arbitrary limits for the wire protocol, but our code was
not checking for them, relying instead on the serialization code to fail
if that happened.

The bad side is if that happened, no information was sent to the host, or
an older version of the information.

This change takes the inverse approach: instead of failing the whole
operation, just ignore any entries that would cause an overflow of the
defined limits. Then at least we provide some data to the host.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>