guest info: use common log domain.
Modify the log calls in the guest info plugin to use the glib functions,
and make sure G_LOG_DOMAIN is set appropriately everywhere by including
the shared header file where it's defined.

The duplicate guestInfoInt.h header was renamed to getlibInt.h, and a
couple of things in it were cleaned up in the process.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>