Text copy using CTRL+A from KDE fails from guests to host
There are 2 different types of selections on Linux: primary and
clipboard. Our internal protocol only supports 1 type at a time. The
reason for the bug is that with current implementation, we almost pick
primary selection if it is available. And in this bug case both primary
and clipboard are available, and clipboard data is more recent, but
we still pick the primary selection. The fix is to first check timestamp,
and if both primary and clipboard are available, we pick the latest one.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>