Cleanup rpcin interface.
Now that vmware-user is gone, there's no code left that needs to work
both with the old RpcIn interface and the new RpcChannel interface. So
we can do some cleanup:

(i) remove the new-style callbacks from the old-style RpcIn interface.
The spots in the old code that used the new API were reverted to use
the old API. So now code using the old RpcIn interface directly is
stuck with the old API.

(ii) Remove the RpcInRet type, which was used by VIX code when it needed
to compile both in vmtoolsd and vmware-user. That type was added in
change 946632.

(iii) Remove some debug logging that was ifdef'd out anyway, and didn't
make sense in this day and age of mostly binary RPCs.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>