Lazily attach console when logging to stdout.
Instead of attaching a console when configuring the logger, do it when
the first log message is written. This ensures that a console only pops
up when the application actually tries to use it. The console is also
detached when the last log domain using it is shut down.

This avoids an issue where consoles would pop up when you configured std
logging for a different application, but all applications sharing the tools
config file ended up attaching a console to their process.

Also fix an issue where an invalid handler was set up for the app's main
log domain, which could cause an ASSERT to be hit.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>