Fix search read error.
Fix a search read issue when we fail to stat an entry but still
want to return success. The entry was supposed to return a fudged
entry with attributes for a regular file, but it failed to set
the attribute mask field to say the type was valid. This causes
the hgfilecopy operation to be canceled.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>