lib/lock: capture the return address of the caller differently
Right now, regardless of builds, the return address of the caller
to an MXUser lock function is captured and passed around. It is only
recorded in one place in a debug build.

Remove the capturing, making things slightly faster in a release
build and provide the ability to see all of the holders of RW lock.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>