VMCI QPair ops must check for valid queues.
When a host (on hosted platforms) creates a VMCI queue pair,
the actual queue content isn't available until a guest
attaches, since the guest is responsible for creating the
queue pair page files and handing them down to the host. Only
at that point will pointers to queue data and queue header be
valid. This works fine with VMCI Sockets in blocking mode,
since connecting a stream socket only returns when both sides
are attached to the queue pair. However, for non-blocking
sockets, we will return earlier - while the stream socket is
still in the connecting state. A subsequent poll to determine
the state of the connecting stream socket may then attempt to
access the queue header to figure out whether the socket has
available data or free space. If the queue header is invalid,
bad things happen. This change adds checks to the VMCIQPair
API implementation, that detects queues with no data, and in
that case reports the queues as having no available data or no
free space, which is essentially true. For the operations that
retrieves pointers, VMCI_ERROR_QUEUEPAIR_NOTATTACHED is now
returned, since we don't want to depict a state that isn't
correct.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>