unity: Clear all windows from tracker on unity exit.
Since we are no longer using the REMOVE_UNTOUCHED flag in our call to
UWT_RequestUpdates(), we must remove any remaining windows left in the
UWT ourselves when we exit unity. Otherwise, if the windows are closed
by the user while we are in single window mode, they are never removed
from the UWT and show up as "blank" rectangles on the host screen the
next time we enter unity.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>