lib/file: File_Rename becomes File_Move
The File_Rename function is really the equivalent of the
unix/Linux "mv" command - it may atomically rename or copy the
file. Unfortunately there is no nice way to know what actually
happened, as cleanup may be necessary upon failure. Fix this.

While we're at it, make the GetLastError/errno make sense.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>