GHI/X11: Fork, reinitialize environment when spawning applications.
Okay, new school GHI leaves a lot of heavy lifting to GLib, because
GLib is awesome.  Unfortunately, its GAppInfo launch handler,
g_app_info_launch, is slightly less awesome in that it doesn't take
an environment vector as an argument.  That makes me sad.

To work around this, I whipped up a wrapper function which does the
following:
- fork()
- reinitialize the environment using clearenv and putenv*
- g_app_info_launch
- Child exits 0 upon success, non-zero on failure.
- Parent blocks until child exits, using child's exit status to
determine success or failure.

* Sort of dirty, but it saves me a parsing step, since nativeEnviron
is composed of strings of key=val, and that's the format accepted by
putenv.  Conversely setenv takes the variable and value strings
separately.

I'm going to file a feature request bug against GIO and provide some
patches.  If accepted upstream, we'll be able to get rid of this
function.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>