GHI/X11: Move "DetectDesktopEnv" to lib/xdg.
This takes the static deskeop environment detection logic out of lib/
ghIntegration and places it a shell script and wrapper in lib/xdg.
We'll now distribute with Tools a vmware-xdg-detect-de which contains
the "detectDE" routine used in many of Portland's xdg-* scripts.

Keeping this logic in a script has a few motivations:
1.  No D-Bus (or similar) dependency added to library code.
2.  Allows users to hack as time goes on in case any of the methods
currently employed stop working.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>