GHI/X11: Add fixed menu item monitors.
This change adds a Gio::FileMonitor to watch over $HOME/.gtk-bookmarks
and connects a slot to Gtk::RecentMgr::signal_changed() to watch for
updates to the recent documents menu.

When an update is detected, the existing relevant submenu is purged
and repopulated.  Then the MenuItemManager emits its menusChanged
signal, alerting the GHI platform code of the change.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>