Move VMCI socket to connected state before sending attach
During VMCI Stream socket connect on Linux, the server side
sends an attach message before adding the pending socket to
the list of connected sockets. External control messages,e.g.,
a WROTE message, triggered by the attach message may arrive
and not find the (previously pending) socket in the list of
connected sockets, and therefore will direct the control
message to the parent socket which is still in the listening
state. In the case of READ/WROTE messages, this will cause the
listening socket to respond with a RST message.

The solution is to add the pending socket to the connected
socket list before sending attach message, since that will
queue incoming packets for the right socket.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>