Introduce VMCI Events for guest memory access
For VMCI users in the vmx, we know generate events when access
to the guest memory changes, i.e., when bus mastering is
enabled/disabled. These events are available through the same
API as the VMCI Event API which is part of the VMCI kernel
API. The events are delivered synchronously, such that actions
can be taken at the point where memory access is
disabled/enabled. The events are delivered with the VMCI
device lock held - the same way that VMCI datagrams are
delivered, so any code that can run in a datagram callback
should be able to run in the event callback as well.

For HGFS, using VMCIEvent_Subscribe during initialization to
subscribe to the event VMCI_EVENT_MEM_ACCESS_OFF will provide
a synchronous callback to the HGFS vmx module whenever the
guest is soft reset or the VMCI driver is unloaded.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>