vmtoolsd: relax VMware hypervisor restriction.
Allow vmtoolsd to run anywhere if not running one of the known Tools
services, or running in debug mode. Also ignore the plugin path not
existing if running in debug mode (so that we can easily run with just
a debug plugin driving everything).

This makes it easier to run vmtoolsd under valgrind, since valgrind
doesn't emulate our backdoor and the code always thinks we're not
running on VMware.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>