Unity/X11: Initialize host:guest desktop mappings before use.
As written there existed a race between entering Unity and receipt
of the host's desired desktop layout such that certain guest X11
events lead to accessing random, uninitialized memory and consequently
receiving ye olde SIGSEGV.

This change works around the issue* by initializing the host:guest
virtual desktop maps based on the guest's current configuration.  That
is, if the guest had 8 desktops, we'll just create an identity mapping
of 8 desktops where desktop 0 in the guest is desktop 0 in the host,
guest1 = host1, and so on.

* Other things should be considered more fully later, such as not even
  registering for events until the host UI issues -ALL- initialization
  RPCs.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>