Fix vsock connect on listening socket
A switch statement in Linux vsock was having a case outside
the enum value range. On closer examination, we are testing
for a value, that is never set. I verified that calling
connect on socket in the listen case will fall through and hit
the default case, since sock->state is still in the
SS_UNCONNECTED state but sk->sk_state is SS_LISTEN. Looking at
other protocol implementations, sk->sk_state is used for the
procotol specific state, so this change simply catches the
SS_LISTEN state in the default case entry, instead of trying
to assign SS_LISTEN to the sock->state when the socket is put
in listening state.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>