vmtoolsd: introduce service properties.
Service properties are a way for plugins to share data without having
link-time dependencies. A plugin can export a property in vmtoolsd's
service object (the same object used currently to send signals), and
that property can be retrieved at any time. Thanks to GObject's property
system, there's also notification of changes, using glib signals.

The properties are not ref counted by the service, so plugins need to agree
on the lifecycle of the property for things to work safely.

Extra: fix an issue in pluginMgr.c where invalid memory was being used
after an array reallocation.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>