Unity/X11: Free desktop mappings when exiting Unity.
When last hacking this code, I added two ASSERTs making sure that
the host:guest desktop map pointers were NULL when entering Unity
and numDesktops == 0.  This is true when first entering Unity
in a single session, but not for subsequent entrances in the same
session.

Rather than keep the mappings until the plugin is destroyed, we'll
now flush them upon exit.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>