Remove unused user space support from VMCI guest drivers
In ancient times, there existed a user level library for
VMCI. It talked to the VMCI guest driver through ioctls. Also,
it needed infrastructure in the guest driver to keep track
of user level processes and their datagram handles.

For all of the modules except windows, we should also be able
to get rid of the character device. I'll do that as a separate
change.

Signed-off-by: Marcelo Vanzin <mvanzin@vmware.com>